vector-app/requirements.txt
pinecone
openai
python-dotenv
